requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart
email-validator
